+++ Shell/KirchhoffLoveShell.C
@@ -5,7 +5,7 @@
 //!
 //! \date Feb 25 2018
 //!
-//! \author ... and ... / NTNU
+//! \author Simen Skogholt Haave and Marit Gaarder Rakvaag / NTNU
 //!
 //! \brief Class for linear Kirchhoff-Love thin shell problems.
 //!
@@ -17,15 +17,7 @@
 #include "ElmMats.h"
 #include "ElmNorm.h"
 #include "Vec3Oper.h"
-#include "Utilities.h"
 #include "IFEM.h"
-
-
-KirchhoffLoveShell::KirchhoffLoveShell () : KirchhoffLove(3)
-{
-  tracFld = nullptr;
-  fluxFld = nullptr;
-}
 
 
 void KirchhoffLoveShell::printLog () const
@@ -62,192 +54,138 @@ bool KirchhoffLoveShell::evalInt (LocalIntegral& elmInt,
 }
 
 
-void KirchhoffLoveShell::formMassMatrix (Matrix& EM, const Vector& N,
-                                         const Vec3& X, double detJW) const
-{
-  double rhow = material->getMassDensity(X)*thickness*detJW;
-  if (rhow == 0.0) return;
-
-  for (size_t a = 1; a <= N.size(); a++)
-    for (size_t b = 1; b <= N.size(); b++)
-      for (unsigned short int i = 1; i <= 3; i++)
-        EM(3*(a-1)+i,3*(b-1)+i) += rhow*N(a)*N(b);
-}
-
-
-void KirchhoffLoveShell::formBodyForce (Vector& ES, const Vector& N, size_t iP,
-                                        const Vec3& X, double detJW) const
-{
-  double p = this->getPressure(X);
-  if (p == 0.0) return;
-
-  for (size_t a = 1; a <= N.size(); a++)
-    ES(3*a) += N(a)*p*detJW;
-
-  // Store pressure value for visualization
-  if (iP < presVal.size())
-    presVal[iP] = std::make_pair(X,Vec3(0.0,0.0,p));
-}
-
-
-Vec3 KirchhoffLoveShell::getTraction (const Vec3& X, const Vec3& n) const
-{
-  if (fluxFld)
-    return (*fluxFld)(X);
-  else if (tracFld)
-    return (*tracFld)(X,n);
-  else
-    return Vec3();
-}
-
-
 bool KirchhoffLoveShell::evalK (Matrix& EK, const FiniteElement& fe,
                                 const Vec3& X) const
 {
+  // Constitutive matrices
+  Matrix Dm, Db;
+  if (!this->formDmatrix(Dm,Db,fe,X))
+    return false;
 
-    int ndof = fe.N.size()*3;
+  // Strain-displacement matrices
+  Matrix Bm, Bb;
+  if (!this->formBmatrix(Bm,Bb,fe))
+    return false;
 
-      Matrix kem(ndof,ndof);
-      Matrix keb(ndof,ndof);
+  Matrix dN_ca, dM_ca;
+  dN_ca.multiply(Dm,Bm); // dN_ca = Dm*Bm, cartesian coordinates
+  dM_ca.multiply(Db,Bb); // dM_ca = Db*Bb, cartesian coordinates
 
-      Matrix dN_ca(3,ndof);
-      Matrix dM_ca(3,ndof);
+  // Element stiffness matrices,
+  // with membrane and bending contributions separated
+  Matrix kem, keb;
+  kem.multiply(Bm,dN_ca,true); // kem = Bm^t*dN_ca = Bm^t*Dm*Bm
+  keb.multiply(Bb,dM_ca,true); // keb = Bb^t*dM_ca = Bb^t*Db*Bb
 
-      Matrix Dm;
-      Matrix Db;
-      if (!this->formDmatrix(Dm,Db,fe,X))
-        return false;
+  EK.add(kem,fe.detJxW).add(keb,fe.detJxW); // EK += (kem+keb)*|J|*w
 
-      Matrix Bm;
-      Matrix Bb;
-      if (!this->formBmatrix(Bm,Bb,fe))
-          return false;
-
-      dN_ca.multiply(Dm,Bm);
-      dM_ca.multiply(Db,Bb);
-
-      kem.multiply(Bm,dN_ca,true,false,true);
-      keb.multiply(Bb,dM_ca,true,false,true);
-      kem.multiply(fe.detJxW);
-      keb.multiply(fe.detJxW);
-
-      EK.add(kem).add(keb);
-
-
-      return true;
-
+  return true;
 }
 
-bool KirchhoffLoveShell::formDmatrix (Matrix& Dm, Matrix& Db, const FiniteElement& fe, const Vec3& X, bool invers) const
+
+bool KirchhoffLoveShell::formDmatrix (Matrix& Dm, Matrix& Db,
+                                      const FiniteElement& fe, const Vec3& X,
+                                      bool invers) const
 {
-    SymmTensor dummy(2); double U;
-    Matrix D;
-    if (!material->evaluate(D,dummy,U,fe,X,dummy,dummy, invers ? -1 : 1))
-        return false;
+  SymmTensor dummy(2); double U;
+  if (!material->evaluate(Dm,dummy,U,fe,X,dummy,dummy, invers ? -1 : 1))
+    return false;
 
-    Matrix Dtemp = D;
-    double factorm = thickness;
-    double factorb = thickness*thickness*thickness/12;
+  double factorm = thickness;
+  double factorb = thickness*thickness*thickness/12.0;
 
-    Dm = D.multiply(invers ? 1.0/factorm : factorm);
-    Db = Dtemp.multiply(invers ? 1.0/factorb : factorb);
+  Db = Dm;
+  Dm.multiply(invers ? 1.0/factorm : factorm);
+  Db.multiply(invers ? 1.0/factorb : factorb);
 
-    return true;
+  return true;
 }
 
-bool KirchhoffLoveShell::formBmatrix (Matrix& Bm, Matrix& Bb, const FiniteElement& fe) const
+
+bool KirchhoffLoveShell::formBmatrix (Matrix& Bm, Matrix& Bb,
+                                      const FiniteElement& fe) const
 {
-    // Covariant basis vectors
-    Vec3 g1 = fe.G.getColumn(1);
-    Vec3 g2 = fe.G.getColumn(2);
+  // Covariant basis vectors
+  Vec3 g1 = fe.G.getColumn(1);
+  Vec3 g2 = fe.G.getColumn(2);
 
-    // Basis vector g3
-    Vec3 g3(g1,g2);
-    double lg3 = g3.length();
-    Vec3 n(g3);
-    n.normalize();
+  // Basis vector g3
+  Vec3 g3(g1,g2);
+  double lg3 = g3.length();
+  Vec3 n(g3);
+  n.normalize();
 
-    // Covariant metric gab
-    double gab11 = g1*g1;
-    double gab12 = g1*g2;
-    double gab22 = g2*g2;
+  // Covariant metric gab
+  double gab11 = g1*g1;
+  double gab12 = g1*g2;
+  double gab22 = g2*g2;
 
-    // Contravariant metric gab_con and base vectors g_con
-    double invdetgab =  1.0/(gab11*gab22-gab12*gab12);
-    double gab_con11 =  invdetgab*gab22;
-    double gab_con12 = -invdetgab*gab12;
-    double gab_con22 =  invdetgab*gab11;
-    Vec3 g1_con = g1*gab_con11 + g2*gab_con12;
-    Vec3 g2_con = g1*gab_con12 + g2*gab_con22;
+  // Contravariant metric gab_con and base vectors g_con
+  double invdetgab =  1.0/(gab11*gab22-gab12*gab12);
+  double gab_con11 =  invdetgab*gab22;
+  double gab_con12 = -invdetgab*gab12;
+  double gab_con22 =  invdetgab*gab11;
+  Vec3 g1_con = g1*gab_con11 + g2*gab_con12;
+  Vec3 g2_con = g1*gab_con12 + g2*gab_con22;
 
-    // Local cartesian coordinates
-    Vec3 e1(g1);     e1.normalize();
-    Vec3 e2(g2_con); e2.normalize();
+  // Local cartesian coordinates
+  Vec3 e1(g1);     e1.normalize();
+  Vec3 e2(g2_con); e2.normalize();
 
-    // Transformation matrix T from contravariant to local cartesian basis
-    Matrix T(3,3);
-    double eg11 = e1*g1_con;
-    double eg12 = e1*g2_con;
-    double eg21 = e2*g1_con;
-    double eg22 = e2*g2_con;
-    T(1,1) =      eg11*eg11;
-    T(1,2) =      eg12*eg12;
-    T(1,3) = 2.0* eg11*eg12;
-    T(2,1) =      eg21*eg21;
-    T(2,2) =      eg22*eg22;
-    T(2,3) = 2.0* eg21*eg22;
-    T(3,1) = 2.0* eg11*eg21;
-    T(3,2) = 2.0* eg12*eg22;
-    T(3,3) = 2.0*(eg11*eg22+eg12*eg21);
+  // Transformation matrix T from contravariant to local cartesian basis
+  Matrix T(3,3);
+  double eg11 = e1*g1_con;
+  double eg12 = e1*g2_con;
+  double eg21 = e2*g1_con;
+  double eg22 = e2*g2_con;
+  T(1,1) =      eg11*eg11;
+  T(1,2) =      eg12*eg12;
+  T(1,3) = 2.0* eg11*eg12;
+  T(2,1) =      eg21*eg21;
+  T(2,2) =      eg22*eg22;
+  T(2,3) = 2.0* eg21*eg22;
+  T(3,1) = 2.0* eg11*eg21;
+  T(3,2) = 2.0* eg12*eg22;
+  T(3,3) = 2.0*(eg11*eg22+eg12*eg21);
 
-    // Strain
-    int ndof = fe.N.size()*3;
-    Matrix dE_ca(3,ndof); // dE_ca = epsilon cartesian coordinates
-    Matrix dK_ca(3,ndof); // dK_ca = kappa cartesian coordinates
-    Matrix dE_cu(3,ndof); // dE_cu = epsilon curvelinear coordinate system
-    Matrix dK_cu(3,ndof); // dK_cu = kappa curvelinear
+  // Strain
+  int ndof = fe.N.size()*3;
+  Matrix dE_cu(3,ndof); // dE_cu = epsilon curvelinear coordinate system
+  Matrix dK_cu(3,ndof); // dK_cu = kappa curvelinear
 
-      for (int i = 1; i <= ndof; i++)
-      {
-        double dummy = i;
-        double k = ceil(dummy/3);
-        int dir = i-3*(k-1);
+    for (int i = 1; i <= ndof; i++)
+    {
+      double dummy = i;
+      double k = ceil(dummy/3);
+      int dir = i-3*(k-1);
 
-        dE_cu(1,i) = fe.dNdX(k,1)*g1(dir);
-        dE_cu(2,i) = fe.dNdX(k,2)*g2(dir);
-        dE_cu(3,i) = 0.5*(fe.dNdX(k,1)*g2(dir) + fe.dNdX(k,2)*g1(dir));
+      dE_cu(1,i) = fe.dNdX(k,1)*g1(dir);
+      dE_cu(2,i) = fe.dNdX(k,2)*g2(dir);
+      dE_cu(3,i) = 0.5*(fe.dNdX(k,1)*g2(dir) + fe.dNdX(k,2)*g1(dir));
 
-        Vec3 dg1;
-        dg1(dir) = fe.dNdX(k,1);
+      Vec3 dg1;
+      dg1(dir) = fe.dNdX(k,1);
 
-        Vec3 dg2;
-        dg2(dir) = fe.dNdX(k,2);
+      Vec3 dg2;
+      dg2(dir) = fe.dNdX(k,2);
 
-        Vec3 dummy2(dg1,g2);
-        Vec3 dg3(g1,dg2);
-        dg3 = dg3 + dummy2;
+      Vec3 dummy2(dg1,g2);
+      Vec3 dg3(g1,dg2);
+      dg3 = dg3 + dummy2;
 
-        double g3dg3lg3_3 = g3*dg3/(lg3*lg3*lg3); // eq 5.31 last part
+      double g3dg3lg3_3 = g3*dg3/(lg3*lg3*lg3);
 
-        Vec3 dn; // eq 5.31 dn = a3,rs (kanskje)
-        for (int kaffi = 1; kaffi <= 3; kaffi ++)
-          {
-             dn(kaffi) = dg3(kaffi)/lg3 - g3(kaffi)*g3dg3lg3_3;
-          }
+      Vec3 dn = dg3/lg3 - g3*g3dg3lg3_3;
 
+      dK_cu(1,i) = -(fe.d2NdX2(k,1,1)*n(dir) + fe.H.getColumn(1)*dn);
+      dK_cu(2,i) = -(fe.d2NdX2(k,2,2)*n(dir) + fe.H.getColumn(2)*dn);
+      dK_cu(3,i) = -(fe.d2NdX2(k,1,2)*n(dir) + fe.H.getColumn(3)*dn);
+    }
 
+  Bm.multiply(T,dE_cu); // Bm = T*dE_cu
+  Bb.multiply(T,dK_cu); // Bb = T*dK_cu
 
-        dK_cu(1,i) = -(fe.d2NdX2(k,1,1)*n(dir) + fe.H.getColumn(1)*dn);
-        dK_cu(2,i) = -(fe.d2NdX2(k,2,2)*n(dir) + fe.H.getColumn(2)*dn);
-        dK_cu(3,i) = -(fe.d2NdX2(k,1,2)*n(dir) + fe.H.getColumn(3)*dn);
-
-
-      } // for int i = 0; i <= ndof; i++
-
-      Bm.multiply(T,dE_cu); // Bm = dE_ca.multiply(T,dE_cu);
-      Bb.multiply(T,dK_cu); // Bb = dK_ca.multiply(T,dK_cu);
-    return true;
-
+  return true;
 }
 
 
@@ -269,40 +207,33 @@ bool KirchhoffLoveShell::evalBou (LocalIntegral& elmInt,
   Vec3 T = this->getTraction(X,normal);
   Vector& ES = static_cast<ElmMats&>(elmInt).b[eS-1];
   for (size_t a = 1; a <= fe.N.size(); a++)
-    for (unsigned short int i = 1; i <= 3; i++)
+    for (int i = 1; i <= 3; i++)
       ES(3*(a-1)+i) += T[i-1]*fe.N(a)*fe.detJxW;
+
+  // Store traction value for visualization
+  if (fe.iGP < tracVal.size() && !T.isZero())
+  {
+    tracVal[fe.iGP].first = X;
+    tracVal[fe.iGP].second += T;
+  }
 
   return true;
 }
 
 
-bool KirchhoffLoveShell::evalSol (Vector& s,
+bool KirchhoffLoveShell::evalSol (Vector& s, const Vector& eV,
                                   const FiniteElement& fe, const Vec3& X,
-                                  const std::vector<int>& MNPC) const
+                                  bool toLocal) const
 {
-  // Extract element displacements
-  Vector eV;
-  if (!primsol.empty() && !primsol.front().empty())
-  {
-    int ierr = utl::gather(MNPC,3,primsol.front(),eV);
-    if (ierr > 0)
-    {
-      std::cerr <<" *** KirchhoffLoveShell::evalSol: Detected "
-                << ierr <<" node numbers out of range."<< std::endl;
-      return false;
-    }
-  }
-  s.reserve(18);
-
-  // Evaluate the stress resultant tensor
-  Vector sb;
-  if (!this->evalSol(s,sb,eV,fe,X,true))
+  // Evaluate the stress resultants (in-plane forces and bending moments)
+  s.reserve(18); Vector sb;
+  if (!this->evalSol(s,sb,eV,fe,X,toLocal))
     return false;
+  else
+    s.insert(s.end(),sb.begin(),sb.end());
 
-  // Append the bending moment to the stress array
-  s.insert(s.end(),sb.begin(),sb.end());
-
-  // Calculate top and bottom surface stresses
+  // Calculate top and bottom surface stresses;
+  // stress tensor components, principal stresses and von Mises stress
   SymmTensor sigma(2);
   Vec3       sigma_p;
   for (int isurf = -1; isurf < 2; isurf += 2)
@@ -350,7 +281,7 @@ bool KirchhoffLoveShell::evalSol (Vector& sm, Vector& sb, const Vector& eV,
   if (!this->formDmatrix(Dm,Db,fe,X))
     return false;
 
-  // Evaluate the membran strain and curvature tensors
+  // Evaluate the membrane strain and curvature tensors
   SymmTensor epsilon(2), kappa(2);
   if (!Bm.multiply(eV,epsilon)) // epsilon = B*eV
     return false;
@@ -464,7 +395,7 @@ bool KirchhoffLoveShellNorm::evalInt (LocalIntegral& elmInt,
     {
       // Evaluate the projected solution
       Vector nr(3), mr(3);
-      for (unsigned short int j = 0; j < 6; j++)
+      for (size_t j = 0; j < 6; j++)
         if (j < 3)
           nr[j] = psol.dot(fe.N,j,nrcmp);
         else
@@ -506,7 +437,7 @@ bool KirchhoffLoveShellNorm::evalBou (LocalIntegral& elmInt,
                                       const Vec3& X, const Vec3& normal) const
 {
   KirchhoffLoveShell& problem = static_cast<KirchhoffLoveShell&>(myProblem);
-  if (!problem.haveLoads()) return true;
+  if (!problem.haveLoads('B')) return true;
 
   ElmNorm& pnorm = static_cast<ElmNorm&>(elmInt);
 
@@ -518,12 +449,6 @@ bool KirchhoffLoveShellNorm::evalBou (LocalIntegral& elmInt,
   // Integrate the external energy
   pnorm[1] += T*u*fe.detJxW;
   return true;
-}
-
-
-int KirchhoffLoveShellNorm::getIntegrandType () const
-{
-  return SECOND_DERIVATIVES;
 }
