+++ Shell/KirchhoffLoveShell.h
@@ -5,7 +5,7 @@
 //!
 //! \date Feb 25 2018
 //!
-//! \author ... and ... / NTNU
+//! \author Simen Skogholt Haave and Marit Gaarder Rakvaag / NTNU
 //!
 //! \brief Class for linear Kirchhoff-Love thin shell problems.
 //!
@@ -28,17 +28,12 @@ class KirchhoffLoveShell : public KirchhoffLove
 {
 public:
   //! \brief Default constructor.
-  KirchhoffLoveShell();
+  KirchhoffLoveShell() : KirchhoffLove(3) {}
   //! \brief Empty destructor,
   virtual ~KirchhoffLoveShell() {}
 
   //! \brief Prints out the problem definition to the log stream.
   virtual void printLog() const;
-
-  //! \brief Defines the traction field to use in Neumann boundary conditions.
-  void setTraction(TractionFunc* tf) { tracFld = tf; }
-  //! \brief Defines the traction field to use in Neumann boundary conditions.
-  void setTraction(VecFunc* tf) { fluxFld = tf; }
 
   using KirchhoffLove::evalInt;
   //! \brief Evaluates the integrand at an interior point.
@@ -58,13 +53,15 @@ public:
                        const Vec3& X, const Vec3& normal) const;
 
   using KirchhoffLove::evalSol;
-  //! \brief Evaluates the secondary solution at a result point.
-  //! \param[out] s Array of solution field values at current point
+  //! \brief Evaluates the finite element (FE) solution at an integration point.
+  //! \param[out] s The FE stress resultant values at current point
+  //! \param[in] eV Element solution vector
   //! \param[in] fe Finite element data at current point
   //! \param[in] X Cartesian coordinates of current point
-  //! \param[in] MNPC Nodal point correspondance for the basis function values
-  virtual bool evalSol(Vector& s, const FiniteElement& fe,
-                       const Vec3& X, const std::vector<int>& MNPC) const;
+  //! \param[in] toLocal If \e true, transform to local coordinates (if defined)
+  virtual bool evalSol(Vector& s, const Vector& eV,
+                       const FiniteElement& fe, const Vec3& X,
+                       bool toLocal = false) const;
 
   //! \brief Evaluates the finite element (FE) solution at an integration point.
   //! \param[out] sm The FE in-plane stress resultant values at current point
@@ -80,29 +77,29 @@ public:
   //! \brief Returns max number of 2ndary solution components to print per line.
   virtual size_t getNo2ndSolPerLine() const { return 6; }
 
-protected:
+private:
   //! \brief Evaluates the stiffness matrix integrand.
+  //! \param EK The element stiffness matrix to receive the contributions
+  //! \param[in] fe Finite element data of current integration point
+  //! \param[in] X Cartesian coordinates of current integration point
   bool evalK(Matrix& EK, const FiniteElement& fe, const Vec3& X) const;
 
-  //! \brief Calculates integration point mass matrix contributions.
-  void formMassMatrix(Matrix& EM, const Vector& N,
-                      const Vec3& X, double detJW) const;
-
-  //! \brief Calculates integration point body force vector contributions.
-  void formBodyForce(Vector& ES, const Vector& N, size_t iP,
-                     const Vec3& X, double detJW) const;
-
   //! \brief Calculates the strain-displacement matrices at current point.
+  //! \param[out] Bm Membrane strain-displacement matrix
+  //! \param[out] Bb Bending strain-displacement matrix
+  //! \param[in] fe Finite element data at current point
   bool formBmatrix(Matrix& Bm, Matrix& Bb, const FiniteElement& fe) const;
 
 public:
   //! \brief Sets up the constitutive matrices at current point.
+  //! \param[out] Dm Constitutive tensor for membrane part
+  //! \param[out] Db Constitutive tensor for bending part
+  //! \param[in] fe Finite element data at current point
+  //! \param[in] X Cartesian coordinates of current point
+  //! \param[in] invers If \e true, the inverse matrices are establised instead
   bool formDmatrix(Matrix& Dm, Matrix& Db,
                    const FiniteElement& fe, const Vec3& X,
                    bool invers = false) const;
-
-  //! \brief Evaluates the boundary traction field (if any) at specified point.
-  Vec3 getTraction(const Vec3& X, const Vec3& n) const;
 
   //! \brief Returns a pointer to an Integrand for solution norm evaluation.
   //! \note The Integrand object is allocated dynamically and has to be deleted
@@ -121,10 +118,6 @@ public:
   //! \param[in] i Field component index
   //! \param[in] prefix Name prefix for all components
   virtual std::string getField2Name(size_t i, const char* prefix) const;
-
-protected:
-  TractionFunc* tracFld; //!< Pointer to implicit boundary traction field
-  VecFunc*      fluxFld; //!< Pointer to explicit boundary traction field
 };
 
 
@@ -159,7 +152,7 @@ public:
                        const Vec3& X, const Vec3& normal) const;
 
   //! \brief Defines which FE quantities are needed by the integrand.
-  virtual int getIntegrandType() const;
+  virtual int getIntegrandType() const { return SECOND_DERIVATIVES; }
 
   //! \brief Returns whether this norm has explicit boundary contributions.
   virtual bool hasBoundaryTerms() const { return true; }
